requirements.txt
amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.5.0
billiard==3.6.4.0
cachetools==5.0.0
celery==4.4.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
Collectfast==2.2.0
cryptography==36.0.1
decorator==5.1.1
Django==3.0.9
django-celery-beat==2.0.0
django-environ==0.4.5
django-extensions==3.0.5
django-filter==2.4.0
django-redis==4.12.1
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.11.1
feedparser==5.2.1
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-secret-manager==2.8.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.0.4
h11==0.9.0
hiredis==1.1.0
httptools==0.1.2
idna==3.3
imgurpython==1.1.7
ipip-ipdb==1.6.0
itsdangerous==0.24
kombu==4.6.11
libcst==0.4.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
Pillow==7.2.0
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==4.0.1
python-telegram-bot==12.8
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tornado==6.1
typing-extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.8
uvicorn==0.11.8
uvloop==0.16.0
uWSGI==2.0.20
vine==1.3.0
websockets==8.1
